requirements.txt
absl-py==1.4.0
asgiref==3.8.1
astroid==2.15.0
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.6
Django==5.0.3
django-model-utils==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna==3.4
inflection==0.5.1
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
protobuf==4.21.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.16.4
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
referencing==0.34.0
requests==2.30.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
thop==0.1.1.post2209072238
tomlkit==0.11.6
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.15.0
